Refactoring for easier WSGI application integration
Simple refactoring to make it easier to deliver the CherryPy
application as a WSGI application that can be used with other
WSGI servers or frameworks such as mod_wsgi. When used as a WSGI
application configuration is done with a YAML file. The default for the
YAML config file is /etc/pyff/pyff_config.yaml but the name of the file
can be set with the process environment variable PYFF_CONFIG.

To separate runtime from configuration the directory to use for the
cache file for remote URL fetching is now configurable. Further
since the WSGI application may then end up in a multi-process environment
(two more more applications running on the same server) the cache file
name is now derived from the PID and a md5 hash of the remote URL.