Fix for issue 142 'Cannot remove a previously selected IdP...'
Fixes issue 142 'Cannot remove a previously selected IdP from discovery'
by fixing an error in the _update method of the discovery_client widget
defined in the file jquery-ds-widget.js. The error was simply the wrong
CSS class assigned as the selector in a call to .on() to bind the click
event.