Fix cleaning of item object
The item object passed around includes an entity property, and that
entity property is itself an object that some templates expect to
contain properties entity_id and entity_icon. This commit moves code for
cleaning the item and making sure there are properties entity_id and
entity_icon into a function.

It also applies that function in the logic branch executed when the
cache_time has expired. This fixes an issue where the ra21.html template
expects entity_id and entity_icon and when they are not present due to
the cache expiring the discovery service stops working.