Prevent error when unpickling RedisWhooshStore from Redis
When Redis is used as the job store the unpickling of an instance of
RedisWhooshStore throws an AttributeError because self._redis does not
exist when the test to see if self._redis is None is executed. This
patch assigns self._redis to getattr(self, '_redis', None) before the
test and prevents the AttributeError exception from being thrown.