Update documentation and examples for content negotiation
Update documentation and examples so that content negotiation
using the "when accept" builtin demonstrates how to configure
the pyFF wsgi app to accept the application/samlmetadata+xml
content type as required by the SAML MDQ specification, but also accept
application/xml so that the service is generous with what it accepts
since not all MDQ clients are compliant. Also change the emit pipe value
to application/samlmetadata+xml so that the documentation helps leads
deployers to configurations that deliver a compliant MDQ service.