Configuration option content_negotation_policy
Added logic to support option content_negotiation_policy. When the
option has value 'extension' then the default is the current behavior:
pyffd will strip off an extension from the path before determining
the entityID and will use the extension to set the return content type
when the accept header is a wildcard or not present. When the option has
value 'adaptive' pyffd will only strip the extension from the path if the
accept header is a wildcard. All other values are the equivalent of
'header' and will be the future default: pyffd will not strip an
extension from the path and will only use the incomming accept header to
determine the return content type.