change to using timezone aware utc_now() functionality in Python 3
All times in SAML are supposed to be "without offset" from UTC, and are
typically expressed as ISO8601 formatted strings with a timezone
component of "Z".

Keep doing that, but use the built-in isoformat() and fromisoformat(),
and thoroughly skip microseconds where we might output such times.

This removes the need for the external dependency 'iso8601'.