Fix type of seed to the random number generator
Since Python 3.11 the seed given to the `random.seed()` function must be
one of the following types: NoneType, int, float, str, bytes, or
bytearray.

Reference: https://docs.python.org/3.11/library/random.html#random.seed

When using `self` as the seed, the random part is the memory address of
the object, which is equivalent to its identity. Using the identity (an
int) should have the same effect in terms of entropy.

Signed-off-by: Ivan Kanakarakis <ivan.kanak@gmail.com>